land packet boundary support

+++ test/test.js
@@ -6,7 +6,6 @@ var spawn = require('child_process').spawn;
 var duplex = require('duplexer');
 
 var createClient = require('net-oce-protocol');
-
 var tdir = __dirname;
 var tmpdir = pjoin(tdir, 'tmp') + '/';
 var exe = pjoin(tdir, '..', 'out/bin/net-oce');
@@ -19,38 +18,52 @@ function setup(fn) {
   });
 
   child.stderr.on('data', function(d) {
-    console.log('ERROR', d.toString());
+    // d.toString().split('\n').forEach(function(line) {
+    //   console.log('>', line);
+    // });
   });
 
-  createClient(duplex(child.stdin, child.stdout), fn);
-
-  child.stdin.write(createClient.request.encode({ method: 0, seq: 0 }));
+  process.nextTick(function() {
+    fn(duplex(child.stdin, child.stdout))
+  });
 
   return child;
 }
 
 function test(name, fn) {
-  var child = setup(function(e, methods) {
-    _test(name, function(t) {
-      var end = t.end.bind(t);
-      var ended = false;
-      t.end = function() {
-        ended = true;
-        child.kill();
-        end();
-      };
+  var child = setup(function(stream) {
+    createClient(stream, function(e, methods) {
+      if (e) throw e;
 
-      child.on('close', function() {
-        if (!ended) {
-          t.fail("net-oce crashed");
-          t.end();
-        }
+      _test(name, function(t) {
+        console.log('d')
+
+        var end = t.end.bind(t);
+        var ended = false;
+        t.end = function() {
+          ended = true;
+          child.kill();
+          end();
+        };
+
+        child.on('close', function() {
+          if (!ended) {
+            t.fail("net-oce crashed");
+            t.end();
+          }
+        });
+
+        fn(methods, t);
       });
 
-      fn(methods, t);
     });
   });
 }
+
+test('test the test', function(methods, t) {
+  t.ok(Object.keys(methods).length);
+  t.end();
+})
 
 test('stl export - no args', function(methods, t) {
   methods.export_stl(function(e, r) {
@@ -61,7 +74,7 @@ test('stl export - no args', function(methods, t) {
 });
 
 test('stl export - invalid SHAPE_HANDLE', function(methods, t) {
-  methods.cube(0, 0, 0, 10, 10, 10, function(e, cube) {
+  methods.prim_cube(10, function(e, cube) {
     methods.export_stl('blah.stl', 0, function(e, r) {
       t.ok(e)
       t.end();
@@ -77,7 +90,7 @@ test('stl export - no shapes', function(methods, t) {
 });
 
 test('stl export - cube', function(methods, t) {
-  methods.cube(0, 0, 0, 10, 10, 10, function(e, cube) {
+  methods.prim_cube(10, function(e, cube) {
 
     t.equal(cube.id, 1);
 
@@ -95,43 +108,53 @@ test('stl export - cube', function(methods, t) {
 });
 
 test('stl export - 2 cubes', function(methods, t) {
-  methods.cube(0, 0, 0, 10, 10, 10, function(e, cube1) {
+  methods.prim_cube(10, function(e, cube1) {
     t.equal(cube1.id, 1);
 
-    methods.cube(20, 0, 0, 10, 10, 10, function(e, cube2) {
-      t.equal(cube2.id, 2);
+    methods.prim_cube(10, function(e, cube2) {
 
-      var out = tmpdir + 'two-cubes.stl';
+      methods.op_translate(20, 0, 0, cube2, function(e, cube3) {
+        t.equal(cube3.id, 3);
 
-      methods.export_stl(out, cube1, cube2, function(e, result) {
-        t.equal(result, true);
+        var out = tmpdir + 'two-cubes.stl';
 
-        var obj = stl.toObject(fs.readFileSync(out).toString());
-        t.equal(obj.facets.length, 24);
+        methods.export_stl(out, cube1, cube2, function(e, result) {
+          t.equal(result, true);
 
-        t.end();
+          var obj = stl.toObject(fs.readFileSync(out).toString());
+          t.equal(obj.facets.length, 24);
+
+          t.end();
+        });
       });
     });
   });
 });
 
 test('cube - no args', function(methods, t) {
-  methods.cube(function(e) {
+  methods.prim_cube(function(e) {
     t.ok(e)
     t.end();
   });
 });
 
 test('cube', function(methods, t) {
-  methods.cube(0, 0, 0, 10, 10, 10, function(e, cube) {
+  methods.prim_cube(10, function(e, cube) {
     t.equal(cube.id, 1);
     t.end();
   });
 });
 
+test('box', function(methods, t) {
+  methods.prim_box(10, 10, 10, function(e, box) {
+    t.equal(box.id, 1);
+    t.end();
+  });
+});
+
 test('op_union - 2 cubes', function(methods, t) {
-  methods.cube(0, 0, 0, 10, 10, 10, function(e, cube1) {
-    methods.cube(5, 5, 5, 10, 10, 10, function(e, cube2) {
+  methods.prim_cube(10, function(e, cube1) {
+    methods.prim_box(20, 10, 10, function(e, cube2) {
       methods.op_union(cube1, cube2, function(e, unioned) {
         t.equal(unioned.id, 3)
 
@@ -141,7 +164,7 @@ test('op_union - 2 cubes', function(methods, t) {
           t.equal(result, true)
 
           var obj = stl.toObject(fs.readFileSync(out).toString());
-          t.equal(obj.facets.length, 36);
+          t.equal(obj.facets.length, 28);
           t.end();
         });
       });
@@ -150,9 +173,9 @@ test('op_union - 2 cubes', function(methods, t) {
 });
 
 test('op_union - 3 cubes', function(methods, t) {
-  methods.cube(0, 0, 0, 10, 10, 10, function(e, cube1) {
-    methods.cube(5, 5, 5, 10, 10, 10, function(e, cube2) {
-      methods.cube(-5, -5, -5, 10, 10, 5, function(e, cube3) {
+  methods.prim_cube(10, function(e, cube1) {
+    methods.prim_box(5, 100, 5, function(e, cube2) {
+      methods.prim_box(100, 5, 5, function(e, cube3) {
         methods.op_union(cube1, cube2, cube3, function(e, unioned) {
 
           var out = tmpdir + 'op_union.3cubes.stl';
@@ -161,7 +184,7 @@ test('op_union - 3 cubes', function(methods, t) {
             t.equal(result, true);
 
             var obj = stl.toObject(fs.readFileSync(out).toString());
-            t.equal(obj.facets.length, 56);
+            t.equal(obj.facets.length, 76);
             t.end();
           });
         });
@@ -170,9 +193,9 @@ test('op_union - 3 cubes', function(methods, t) {
   });
 });
 
-test('op_union - invalid handle in loop', function(methods, t) {
-  methods.cube(0, 0, 0, 10, 10, 10, function(e, cube1) {
-    methods.cube(5, 5, 5, 10, 10, 10, function(e, cube2) {
+test('op_union - invalid handle', function(methods, t) {
+  methods.prim_box(100, 5, 10, function(e, cube1) {
+    methods.prim_cube(10, function(e, cube2) {
       methods.op_union(cube1, cube2, { id: 0 }, function(e, r) {
         t.ok(e)
         t.end();
@@ -181,9 +204,9 @@ test('op_union - invalid handle in loop', function(methods, t) {
   });
 });
 
-test('op_union - invalid handle in loop', function(methods, t) {
-  methods.cube(0, 0, 0, 10, 10, 10, function(e, cube1) {
-    methods.cube(5, 5, 5, 10, 10, 10, function(e, cube2) {
+test('op_union - invalid handle', function(methods, t) {
+  methods.prim_cube(10, function(e, cube1) {
+    methods.prim_box(100, 5, 5, function(e, cube2) {
       methods.op_union(cube1, cube2, 0, function(e, r) {
         t.ok(e);
         t.end();
@@ -192,9 +215,9 @@ test('op_union - invalid handle in loop', function(methods, t) {
   });
 });
 
-test('op_cut - two cubes', function(methods, t) {
-  methods.cube(0, 0, 0, 20, 20, 10, function(e, cube1) {
-    methods.cube(10, 10, 0, 5, 5, 20, function(e, cube2) {
+test('op_cut - two boxes', function(methods, t) {
+  methods.prim_box(20, 20, 10, function(e, cube1) {
+    methods.prim_box(5, 5, 20, function(e, cube2) {
       methods.op_cut(cube1, cube2, function(e, cut) {
         t.ok(cut.id);
 
@@ -209,5 +232,47 @@ test('op_cut - two cubes', function(methods, t) {
         });
       });
     });
+  });
+});
+
+test('cube - 10 in one tick', function(methods, t) {
+  var total = 10;
+  var seen = 0;
+  function after(e, r) {
+    t.ok(!e && r);
+    seen++;
+    console.log(seen, 'vs', total);
+    if (seen === total) {
+
+      t.end();
+    }
+  }
+
+  for (var i = 0; i<total; i++) {
+    methods.prim_cube(10 * (i + 1), after);
+  }
+});
+
+_test('partial length message', function(t) {
+  var child = setup(function(stream) {
+
+    stream.once('data', function(d) {
+      t.ok(d);
+      child.kill();
+      t.end();
+    });
+
+    var buffer = createClient.encodeRequest({ method: 0, seq: 0 });
+    var where = 0;
+    setTimeout(function write() {
+      var slice = buffer.slice(where, where+1);
+      stream.write(slice);
+      console.log(where, slice.length, buffer.slice(where).length)
+      where++;
+
+      if (where < buffer.length) {
+        setTimeout(write, 16);
+      }
+    }, 100);
   });
 });
